requirements.txt
opencv-python
torch
torchvision
numpy
Pillow
flask
websocket-client
joblib
pandas
urllib3
